error

+++ src/main/java/edu/psu/os/KV511/model/Message.java
@@ -7,10 +7,10 @@ public class Message implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	private String instruction;
-	private long key;
-	private long value;
+	private int key;
+	private int value;
 	
-	public Message(String instruction, long key, long value) {
+	public Message(String instruction, int key, int value) {
 		super();
 		this.instruction = instruction;
 		this.key = key;
@@ -26,13 +26,13 @@ public class Message implements Serializable {
 	public long getKey() {
 		return key;
 	}
-	public void setKey(long key) {
+	public void setKey(int key) {
 		this.key = key;
 	}
 	public long getValue() {
 		return value;
 	}
-	public void setValue(long value) {
+	public void setValue(int value) {
 		this.value = value;
 	}
